[DEV] ✨🐛 Finished fixing (un)liking bugs

+++ src/_11_e2ecrudapp/pages/home/PostInfo.tsx
@@ -3,6 +3,7 @@ import { PostInfoType } from "./Home";
 
 // Hooks imports
 import { useEffect, useState } from "react";
+import { useNavigate } from "react-router-dom";
 
 // Firestore database imports
 import { dbDriver } from "../../config/firebaseActual";
@@ -26,19 +27,23 @@ export const PostInfo = ({ postInfo }: Args): JSX.Element => {
   // Current user
   const [currentUser] = useAuthState(auth);
 
+  // Navigate hooks (Navigate to login if not connected)
+  const navigate = useNavigate();
+
   // Like hooks initialization
   const [likes, setLikes] = useState<Like[]>([]);
   useEffect(() => {
     // Get data from DB
     const getLikesHdle = async () => {
-      const data = await getDocs(collection(dbDriver, "likes"));
+      const likeOfThisPostQuery = query(collection(dbDriver, "likes"), where("postId", "==", postInfo.postId));
+      const data = await getDocs(likeOfThisPostQuery);
       const formattedData = data.docs.map((doc) => doc.data() as Like);
       setLikes(formattedData);
     };
 
     // Launch get
     getLikesHdle();
-  }, []);
+  }, [postInfo]);
 
   // Likes info initialization
   const nbLikes = likes.length;
@@ -46,50 +51,64 @@ export const PostInfo = ({ postInfo }: Args): JSX.Element => {
 
   // Handler for liking
   const hdleLike = async () => {
-    try {
-      // Launch DB add like
-      const like = { postId: postInfo.postId, userId: currentUser?.uid } as Like;
-      await addDoc(collection(dbDriver, "likes"), like);
-
-      // Update page
-      setLikes([...likes, like]);
-    } catch (err) {
-      alert(err);
+    // If not connected go to login instead
+    if (!currentUser) {
+      alert("Please login first 😉");
+      navigate("/login");
     }
+    // Else manage liking
+    else
+      try {
+        // Launch DB add like
+        const like: Like = { postId: postInfo.postId, userId: currentUser?.uid };
+        await addDoc(collection(dbDriver, "likes"), like);
+
+        // Update page
+        setLikes([...likes, like]);
+      } catch (err) {
+        alert(err);
+      }
   };
 
   // Handler for unliking
   const hdleUnlike = async () => {
-    try {
-      // Retrieve Ids to delete
-      const likeToDeleteQuery = query(
-        collection(dbDriver, "likes"),
-        where("postId", "==", postInfo.postId),
-        where("userId", "==", currentUser?.uid)
-      );
-      const likeToDeleteId = (await getDocs(likeToDeleteQuery)).docs[0].id; // Suppose one element otherwise exception
-
-      // Launch delete
-      const likeToDeleteDoc = doc(dbDriver, "likes", likeToDeleteId);
-      await deleteDoc(likeToDeleteDoc);
-
-      // Update page
-      const upToDateLike = likes.filter((x) => x.postId === postInfo.postId && x.userId === currentUser?.uid);
-      setLikes(upToDateLike);
-    } catch (err) {
-      alert(err);
+    // If not connected go to login instead
+    if (!currentUser) {
+      alert("Please login first 😉");
+      navigate("/login");
     }
+    // Else manage liking
+    else
+      try {
+        // Retrieve Ids to delete
+        const likeToDeleteQuery = query(
+          collection(dbDriver, "likes"),
+          where("postId", "==", postInfo.postId),
+          where("userId", "==", currentUser?.uid)
+        );
+        const likeToDeleteId = (await getDocs(likeToDeleteQuery)).docs[0].id; // Suppose one element otherwise corrupted
+
+        // Launch delete
+        const likeToDeleteDoc = doc(dbDriver, "likes", likeToDeleteId);
+        await deleteDoc(likeToDeleteDoc);
+
+        // Update page
+        const upToDateLike = likes.filter((x) => !(x.postId === postInfo.postId && x.userId === currentUser?.uid));
+        setLikes(upToDateLike);
+      } catch (err) {
+        alert(err);
+      }
   };
 
   // Render
   return (
-    <div>
+    <>
       <h1>{postInfo.title}</h1>
       <p>{postInfo.description}</p>
       <p>@ {postInfo.userName}</p>
       <p>At {postInfo.createdAt}</p>
       {isAlreadyLiked ? <button onClick={hdleLike}>👍</button> : <button onClick={hdleUnlike}>👎</button>}
       {nbLikes !== 0 && <p>Likes: {nbLikes}</p>}
-    </div>
+    </>
   );
 };
